Validate IP addresses upon opening new sockets
Fixes #2.

Instead of rewriting the hostname to be the IP address, instead we
look at the resolved IP just before cURL opens the socket. This is
fine since cURL's connection pools are per-handle so we don't have
to worry about keep-alive and other handles' connections.

This negates the need for DNS pinning, so enabling DNS pinning is
now effectively a no-op.